server/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
sqlmodel==0.0.22
pydantic-settings==2.7.1
httpx==0.28.1
pillow==11.1.0
pytest==8.3.4
pytest-asyncio==0.25.2
